Add: Tree API 작업

+++ backend-2/src/main/java/io/ggogit/ggogit/api/tree/dto/BookTreeResponse.java
@@ -32,7 +32,7 @@ public class BookTreeResponse {
     private String description;
     private Boolean visibility;
     private String createdAt;
-    private Long readingPage;
+    private Integer readingPage;
     private String coverImageName;
 
     public static BookTreeResponse of(Tree tree) {
